+++ src/android/BackgroundGpsPlugin.java
@@ -32,6 +32,11 @@ import org.json.JSONObject;
 import com.marianhello.cordova.bgloc.Config;
 import com.marianhello.cordova.bgloc.Constant;
 import com.marianhello.cordova.bgloc.ServiceProvider;
+import com.tenforwardconsulting.cordova.bgloc.data.LocationDAO;
+import com.tenforwardconsulting.cordova.bgloc.data.DAOFactory;
+import com.tenforwardconsulting.cordova.bgloc.data.LocationProxy;
+
+import java.util.Collection;
 
 public class BackgroundGpsPlugin extends CordovaPlugin {
     private static final String TAG = "BackgroundGpsPlugin";
@@ -44,6 +49,9 @@ public class BackgroundGpsPlugin extends CordovaPlugin {
     public static final String ACTION_SHOW_LOCATION_SETTINGS = "showLocationSettings";
     public static final String REGISTER_MODE_CHANGED_RECEIVER = "watchLocationMode";
     public static final String UNREGISTER_MODE_CHANGED_RECEIVER = "stopWatchingLocationMode";
+    public static final String ACTION_GET_ALL_LOCATIONS = "getLocations";
+    public static final String ACTION_DELETE_LOCATION = "deleteLocation";
+    public static final String ACTION_DELETE_ALL_LOCATIONS = "deleteAllLocations";
 
     private Config config = new Config();
     private Boolean isEnabled = false;
@@ -51,11 +59,29 @@ public class BackgroundGpsPlugin extends CordovaPlugin {
     private CallbackContext callbackContext;
     private CallbackContext locationModeChangeCallbackContext;
 
-    private BroadcastReceiver mMessageReceiver = new BroadcastReceiver() {
+    private BroadcastReceiver actionReceiver = new BroadcastReceiver() {
         @Override
         public void onReceive(Context context, Intent intent) {
             Log.d(TAG, "Received location from bg service");
-            handleMessage(intent);
+            Bundle data = intent.getExtras();
+            switch (data.getInt(Constant.ACTION)) {
+                case Constant.ACTION_LOCATION_UPDATE:
+                    try {
+                        JSONObject location = new JSONObject(data.getString(Constant.DATA));
+                        PluginResult result = new PluginResult(PluginResult.Status.OK, location);
+                        result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(result);
+                        Log.d(TAG, "Sending plugin result");
+                    } catch (JSONException e) {
+                        PluginResult result = new PluginResult(PluginResult.Status.JSON_EXCEPTION);
+                        result.setKeepCallback(true);
+                        callbackContext.sendPluginResult(result);
+                        Log.w(TAG, "Error converting message to json");
+                    }
+                    break;
+                default:
+                    break;
+            }
         }
     };
 
@@ -82,16 +108,16 @@ public class BackgroundGpsPlugin extends CordovaPlugin {
         Activity activity = this.cordova.getActivity();
         Context context = activity.getApplicationContext();
 
-        if (ACTION_START.equalsIgnoreCase(action) && !isEnabled) {
+        if (ACTION_START.equals(action) && !isEnabled) {
             try {
-                updateServiceIntent = new Intent(activity, ServiceProvider.getClass(config.getLocationServiceProvider()));
+                updateServiceIntent = new Intent(activity, ServiceProvider.getClass(config.getServiceProvider()));
             } catch (ClassNotFoundException e) {
                 callbackContext.error("Configuration error: provider not found");
                 return false;
             }
 
-            IntentFilter intentFilter = new IntentFilter(Constant.FILTER);
-            context.registerReceiver(mMessageReceiver, intentFilter);
+            IntentFilter intentFilter = new IntentFilter(Constant.ACTION_FILTER);
+            context.registerReceiver(actionReceiver, intentFilter);
             String canonicalName = activity.getClass().getCanonicalName();
 
             updateServiceIntent.putExtra("config", config);
@@ -103,26 +129,27 @@ public class BackgroundGpsPlugin extends CordovaPlugin {
             isEnabled = true;
             Log.d(TAG, "bg service has been started");
 
-        } else if (ACTION_STOP.equalsIgnoreCase(action)) {
-            context.unregisterReceiver(mMessageReceiver);
+        } else if (ACTION_STOP.equals(action)) {
+            context.unregisterReceiver(actionReceiver);
             isEnabled = false;
             activity.stopService(updateServiceIntent);
             callbackContext.success();
             Log.d(TAG, "bg service has been stopped");
-        } else if (ACTION_CONFIGURE.equalsIgnoreCase(action)) {
+        } else if (ACTION_CONFIGURE.equals(action)) {
             try {
                 this.callbackContext = callbackContext;
                 this.config = Config.fromJSONArray(data);
                 Log.d(TAG, "bg service configured");
+                // callbackContext.success(); //we cannot do this
             } catch (JSONException e) {
                 callbackContext.error("Configuration error: " + e.getMessage());
                 return false;
             }
-        } else if (ACTION_SET_CONFIG.equalsIgnoreCase(action)) {
+        } else if (ACTION_SET_CONFIG.equals(action)) {
             // TODO reconfigure Service
             callbackContext.success();
             Log.d(TAG, "bg service reconfigured");
-        } else if (ACTION_LOCATION_ENABLED_CHECK.equalsIgnoreCase(action)) {
+        } else if (ACTION_LOCATION_ENABLED_CHECK.equals(action)) {
             Log.d(TAG, "location services enabled check");
             try {
                 int isLocationEnabled = BackgroundGpsPlugin.isLocationEnabled(context) ? 1 : 0;
@@ -136,10 +163,26 @@ public class BackgroundGpsPlugin extends CordovaPlugin {
         } else if (REGISTER_MODE_CHANGED_RECEIVER.equals(action)) {
             this.locationModeChangeCallbackContext = callbackContext;
             context.registerReceiver(locationModeChangeReceiver, new IntentFilter(LocationManager.MODE_CHANGED_ACTION));
-
         } else if (UNREGISTER_MODE_CHANGED_RECEIVER.equals(action)) {
             context.unregisterReceiver(locationModeChangeReceiver);
             this.locationModeChangeCallbackContext = null;
+        } else if (ACTION_GET_ALL_LOCATIONS.equals(action)) {
+            try {
+                callbackContext.success(this.getAllLocations());
+            } catch (JSONException e) {
+                callbackContext.error("Converting locations to JSON failed.");
+            }
+        } else if (ACTION_DELETE_LOCATION.equals(action)) {
+            try {
+                this.deleteLocation(data.getInt(0));
+                callbackContext.success();
+            } catch (JSONException e) {
+                callbackContext.error("Configuration error: " + e.getMessage());
+                return false;
+            }
+        } else if (ACTION_DELETE_ALL_LOCATIONS.equals(action)) {
+            this.deleteAllLocations();
+            callbackContext.success();
         }
 
         return true;
@@ -150,10 +193,19 @@ public class BackgroundGpsPlugin extends CordovaPlugin {
      * Checks to see if it should turn off
      */
     public void onDestroy() {
+        Log.d(TAG, "Main Activity destroyed!!!");
         Activity activity = this.cordova.getActivity();
 
-        if (isEnabled && config.getStopOnTerminate()) {
-            activity.stopService(updateServiceIntent);
+        if (isEnabled) {
+            if (config.getStopOnTerminate()) {
+                activity.stopService(updateServiceIntent);
+            } else {
+                //todo: send info to location service
+                Intent intent = new Intent(Constant.ACTION_FILTER);
+                intent.putExtra(Constant.ACTION, Constant.ACTION_ACTIVITY_KILLED);
+                intent.putExtra(Constant.DATA, true);
+                activity.sendBroadcast(intent);
+            }
         }
     }
 
@@ -176,23 +228,26 @@ public class BackgroundGpsPlugin extends CordovaPlugin {
         }
     }
 
-    private void handleMessage(Intent msg) {
-        Bundle data = msg.getExtras();
-        switch (data.getInt(Constant.COMMAND, 0))
-        {
-            case Constant.UPDATE_PROGRESS:
-                try {
-                    JSONObject location = new JSONObject(data.getString(Constant.DATA));
-                    PluginResult result = new PluginResult(PluginResult.Status.OK, location);
-                    result.setKeepCallback(true);
-                    this.callbackContext.sendPluginResult(result);
-                    Log.d(TAG, "Sending plugin result");
-                } catch (JSONException e) {
-                    Log.w(TAG, "Error converting message to json");
-                }
-                break;
-            default:
-                break;
+    public JSONArray getAllLocations() throws JSONException {
+        JSONArray jsonLocationsArray = new JSONArray();
+        Context context = this.cordova.getActivity().getApplicationContext();
+        LocationDAO dao = DAOFactory.createLocationDAO(context);
+        Collection<LocationProxy> locations = dao.getAllLocations();
+        for (LocationProxy location : locations) {
+            jsonLocationsArray.put(location.toJSONObject());
         }
+        return jsonLocationsArray;
+    }
+
+    public void deleteLocation(Integer locationId) {
+        Context context = this.cordova.getActivity().getApplicationContext();
+        LocationDAO dao = DAOFactory.createLocationDAO(context);
+        dao.deleteLocation(locationId);
+    }
+
+    public void deleteAllLocations() {
+        Context context = this.cordova.getActivity().getApplicationContext();
+        LocationDAO dao = DAOFactory.createLocationDAO(context);
+        dao.deleteAllLocations();
     }
 }