+++ src/android/Constant.java
@@ -14,12 +14,12 @@ package com.marianhello.cordova.bgloc;
  */
 public abstract class Constant
 {
-    public static final int UPDATE_PROGRESS = 1;
-    public static final int RESULT = 2;
     public static final String DATA = "DATA";
-    public static final String COMMAND = "COMMAND";
-    public static final String FILTER = "com.marianhello.cordova.bgloc.communication.REQUEST_PROCESSED";
-    public static final String LOCATION_UPDATE = "com.tenforwardconsulting.cordova.bgloc.LOCATION_UPDATE";
-    public static final String STOP_RECORDING = "com.tenforwardconsulting.cordova.bgloc.STOP_RECORDING";
-    public static final String START_RECORDING = "com.tenforwardconsulting.cordova.bgloc.START_RECORDING";
+    public static final String ACTION = "ACTION";
+    public static final String ACTION_FILTER = "com.tenforwardconsulting.cordova.bgloc.ACTION";
+    public static final String LOCATION_UPDATE_FILTER = "com.tenforwardconsulting.cordova.bgloc.LOCATION_UPDATE";
+    public static final int ACTION_LOCATION_UPDATE = 0;
+    public static final int ACTION_STOP_RECORDING = 1;
+    public static final int ACTION_START_RECORDING = 2;
+    public static final int ACTION_ACTIVITY_KILLED = 3;
 }